projects/03_coffee_shop_full_stack/starter_code/backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
SQLAlchemy==1.4.20
python-jose==3.3.0